Note dictionary updates from Vicky
I took out the few h for half notes so all we have are eighth and
quarter notes.  I think I matched up the reverse codon part as well.

The last 2 note changes I made may not follow the wobble rule exactly
but we had to do something to fill in the gaps for the missing high A
and high B notes